# storage: add prometheus to memory implementation