The simplest statistics for the memory implementation.

I wonder if it'd be better to record these asynchronously in a designated goroutine?

Benchmark changes:
```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkPut                              350           347           -0.86%
BenchmarkPut-8                            565           550           -2.65%
BenchmarkPut1k                            379           386           +1.85%
BenchmarkPut1k-8                          602           627           +4.15%
BenchmarkPut1kInfohash                    397           404           +1.76%
BenchmarkPut1kInfohash-8                  106           114           +7.55%
BenchmarkPut1kInfohash1k                  400           416           +4.00%
BenchmarkPut1kInfohash1k-8                112           112           +0.00%
BenchmarkPutDelete                        1267          1280          +1.03%
BenchmarkPutDelete-8                      3278          3126          -4.64%
BenchmarkPutDelete1k                      1254          1284          +2.39%
BenchmarkPutDelete1k-8                    3423          3083          -9.93%
BenchmarkPutDelete1kInfohash              1303          1345          +3.22%
BenchmarkPutDelete1kInfohash-8            2939          3304          +12.42%
BenchmarkPutDelete1kInfohash1k            1381          1362          -1.38%
BenchmarkPutDelete1kInfohash1k-8          3216          3355          +4.32%
BenchmarkDeleteNonexist                   222           210           -5.41%
BenchmarkDeleteNonexist-8                 257           256           -0.39%
BenchmarkDeleteNonexist1k                 216           217           +0.46%
BenchmarkDeleteNonexist1k-8               257           256           -0.39%
BenchmarkDeleteNonexist1kInfohash         226           222           -1.77%
BenchmarkDeleteNonexist1kInfohash-8       64.6          64.2          -0.62%
BenchmarkDeleteNonexist1kInfohash1k       227           220           -3.08%
BenchmarkDeleteNonexist1kInfohash1k-8     65.3          65.3          +0.00%
BenchmarkPutGradDelete                    1922          1931          +0.47%
BenchmarkPutGradDelete-8                  4490          2838          -36.79%
BenchmarkPutGradDelete1k                  1854          2051          +10.63%
BenchmarkPutGradDelete1k-8                4439          2795          -37.04%
BenchmarkPutGradDelete1kInfohash          1954          2222          +13.72%
BenchmarkPutGradDelete1kInfohash-8        4699          3742          -20.37%
BenchmarkPutGradDelete1kInfohash1k        1968          2021          +2.69%
BenchmarkPutGradDelete1kInfohash1k-8      4634          4304          -7.12%
BenchmarkGradNonexist                     389           384           -1.29%
BenchmarkGradNonexist-8                   656           622           -5.18%
BenchmarkGradNonexist1k                   420           414           -1.43%
BenchmarkGradNonexist1k-8                 716           700           -2.23%
BenchmarkGradNonexist1kInfohash           441           444           +0.68%
BenchmarkGradNonexist1kInfohash-8         119           120           +0.84%
BenchmarkGradNonexist1kInfohash1k         436           439           +0.69%
BenchmarkGradNonexist1kInfohash1k-8       119           124           +4.20%
BenchmarkAnnounceLeecher                  17939         18179         +1.34%
BenchmarkAnnounceLeecher-8                4193          4171          -0.52%
BenchmarkAnnounceLeecher1kInfohash        22033         21650         -1.74%
BenchmarkAnnounceLeecher1kInfohash-8      5060          5031          -0.57%
BenchmarkAnnounceSeeder                   18580         17771         -4.35%
BenchmarkAnnounceSeeder-8                 4281          4226          -1.28%
BenchmarkAnnounceSeeder1kInfohash         22403         21696         -3.16%
BenchmarkAnnounceSeeder1kInfohash-8       5045          5019          -0.52%
```

Some stuff goes up, some stuff goes down(?). Maybe just jitter?